+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/agent/task/meta/PipeTemporaryMeta.java
@@ -19,75 +19,4 @@
 
 package org.apache.iotdb.commons.pipe.agent.task.meta;
 
-import java.util.Collections;
-import java.util.Objects;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-
-public class PipeTemporaryMeta {
-
-  private final Set<Integer> completedDataNodeIds =
-      Collections.newSetFromMap(new ConcurrentHashMap<>());
-  private final ConcurrentMap<Integer, Long> nodeId2RemainingEventMap = new ConcurrentHashMap<>();
-  private final ConcurrentMap<Integer, Double> nodeId2RemainingTimeMap = new ConcurrentHashMap<>();
-
-  public void markDataNodeCompleted(final int dataNodeId) {
-    completedDataNodeIds.add(dataNodeId);
-  }
-
-  public void markDataNodeUncompleted(final int dataNodeId) {
-    completedDataNodeIds.remove(dataNodeId);
-  }
-
-  public void setRemainingEvent(final int dataNodeId, final long remainingEventCount) {
-    nodeId2RemainingEventMap.put(dataNodeId, remainingEventCount);
-  }
-
-  public void setRemainingTime(final int dataNodeId, final double remainingTime) {
-    nodeId2RemainingTimeMap.put(dataNodeId, remainingTime);
-  }
-
-  public Set<Integer> getCompletedDataNodeIds() {
-    return completedDataNodeIds;
-  }
-
-  public long getGlobalRemainingEvents() {
-    return nodeId2RemainingEventMap.values().stream().reduce(Long::sum).orElse(0L);
-  }
-
-  public double getGlobalRemainingTime() {
-    return nodeId2RemainingTimeMap.values().stream().reduce(Math::max).orElse(0d);
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    final PipeTemporaryMeta that = (PipeTemporaryMeta) o;
-    return Objects.equals(this.completedDataNodeIds, that.completedDataNodeIds)
-        && Objects.equals(this.nodeId2RemainingEventMap, that.nodeId2RemainingEventMap)
-        && Objects.equals(this.nodeId2RemainingTimeMap, that.nodeId2RemainingTimeMap);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(completedDataNodeIds, nodeId2RemainingEventMap, nodeId2RemainingTimeMap);
-  }
-
-  @Override
-  public String toString() {
-    return "PipeTemporaryMeta{"
-        + "completedDataNodeIds="
-        + completedDataNodeIds
-        + ", nodeId2RemainingEventMap="
-        + nodeId2RemainingEventMap
-        + ", nodeId2RemainingTimeMap"
-        + nodeId2RemainingTimeMap
-        + '}';
-  }
-}
+public interface PipeTemporaryMeta {}
